Refactored the skeleton Java code.
Moved all classes from `Type.java` into their own `type package.

Moved all classes from `Ast.java` into their own `ast` package. They
are also prefixed with `Ast` now. Part of the reason is that there is
a `Float` class in there...

Changed the visibility of the functions involved (from package-local,
to public), to preserve the functionality of the code after the above
modification.

Added a `final` modifier to the fields of the AST classes and to
most of the classes, as needed.

Finally, the code looks more like Java code.