Made sure that `mvn package` produces a standalone, executable jar.
Got rid of the `java-cup-11b-runtime.jar` because Maven kept complaining about
<scope>system</scope> and because the `java_cup.runtime` package wasn't
automatically included in the final JAR (and it would have been a pain to do so).

Because I couldn't find the sources for `java-cup-11b-runtime.jar`, but I found
them for version 11a, I switched the CUP compiler to 11a and dumped the
`java_cup.runtime` sources that I found into `src/main/java` (so they will get
included in the JAR).

I know this is not beautiful, but it seems somewhat sane compared to other options.