Replaced CUP with a modified version (made by me in the past).
That sorry excuse of a parser generator caused me a lot of hassle in the past
and I needed it to (1) generate code slightly differently and (2) implement the
runtime library differently.

This is why, as part of a project I did 1 year ago, I forked the implementation
of CUP, moved its code to a whole new package (`ldf`) to avoid conflicts, and
then changed its code generation and runtime library as I wished. This was
especially to give me access to the `java_cup.runtime.Symbol` objects, which
would allow me to recover information about the tokens and their position in
the source code.

Now I know the exact position in the source code (line and column) of each AST
node. For example, for every identifier, I know where it was declared, so that
I can later implement error reporting to point to the source of an error.

Other changes:
- Moved `Id` into the `ast` package, because that's where I think it belongs.
- Created `AstNode` to carry information about a node's location in the input.
- Changed `Id` and `AstExp` to extend `AstNode`.