Added support for global variables in `FreeVarVisitor`.
With this, some predefined symbols (such as `print_int`) will not be
reported as "free" just because they are not bound within the

Also, integrated it with the compiler, to check for free variables and
report them as errors.

Added unit test to run all files in `src/test/resources/mini_camel/tests`
through this checker. No free variables must exist, except for a list of
predefined functions.